@@ -7,12 +7,11 @@
     "build": "vue-cli-service build"
   },
   "dependencies": {
-    "@braid/vue-formulate": "^2.5.3",
-    "bootstrap": "^5.3.0-alpha1",
+    "@formkit/vue": "^0.16.1",
     "core-js": "^3.8.3",
-    "vue": "^2.6.14",
-    "vue-router": "^3.5.1",
-    "vuex": "^3.6.2",
+    "vue": "^3.2.13",
+    "vue-router": "^4.0.13",
+    "vuex": "^4.0.0",
     "vuex-persist": "^3.1.3"
   },
   "devDependencies": {
