@@ -1,16 +1,8 @@
-import Vue from 'vue'
+import { createApp } from 'vue'
 import App from './App.vue'
 import router from './router'
 import store from './store'
+import { plugin, defaultConfig } from '@formkit/vue'
 import './styles.css'
-const VueFormulate = require('@braid/vue-formulate')
 
-
-Vue.config.productionTip = false
-Vue.use(VueFormulate.default)
-
-new Vue({
-  router,
-  store,
-  render: h => h(App)
-}).$mount('#app')
+createApp(App).use(store).use(router).use(plugin, defaultConfig).mount('#app')
