@@ -23,7 +23,10 @@ export default {
         font-family: Avenir, Helvetica, Arial, sans-serif;
         -webkit-font-smoothing: antialiased;
         -moz-osx-font-smoothing: grayscale;
-        text-align: center;
+        max-width: 940px;
+        max-height: 600px;
+        height: 100%;
+        width: 100%;
         color: #2c3e50;
     }
 
